+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.10.1 - 2021-08-21
 
 * Fix generation when source/target class is a `public` class enclosed an `internal` class
+* Add new combinators: `mapError`, `bind` (up to 14 arguments)
 
 ## 0.10.0 - 2021-08-06